+++ assets/javascripts/kitten/components/accordeon/components/header.js
@@ -1,0 +1,34 @@
+import React from 'react'
+import styled, { css } from 'styled-components'
+import COLORS from '../../../constants/colors-config'
+import { pxToRem } from '../../../helpers/utils/typography'
+import { Context } from './context'
+
+const HeaderStyled = styled.div`
+  box-sizing: border-box;
+  padding: ${pxToRem(20)};
+  min-height: ${pxToRem(30)};
+  border: ${pxToRem(2)} solid ${COLORS.line1};
+  border-radius: ${pxToRem(3)};
+  cursor: pointer;
+
+  ${({ isSelected }) =>
+    isSelected &&
+    css`
+      border-bottom-left-radius: 0;
+      border-bottom-right-radius: 0;
+    `}
+`
+
+export const Header = ({ id, children }) => (
+  <Context.Consumer>
+    {({ updateSelectedItem, selectedItem }) => (
+      <HeaderStyled
+        isSelected={selectedItem === id}
+        onClick={() => updateSelectedItem(id)}
+      >
+        {children}
+      </HeaderStyled>
+    )}
+  </Context.Consumer>
+)