+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `Accordeon` component.
+- Feature: Add `getReactElementsByType` helper.
+- Feature: Uses `style` prop on `RadioButton`.
+
 ## [1.1.3] - 2018-12-20
 
 Fix: